Refactored to stop using location reload and start retreving single updated item. Changed order item

+++ FoodDelivery.DAL/EF/Entities/Restaurant.cs
@@ -17,7 +17,7 @@ namespace FoodDelivery.DAL.EF.Entities
 
 		public string Description { get; set; }
 
-		public double? Rating { get; set; }
+		public decimal? Rating { get; set; }
 		public int Status { get; set; }
 
 		[ForeignKey("Type")]
